Rework back-office routes, sidebar & proxy
Move many back-office pages under /back-office, update internal links and route pushes across components (appointments, users, customers, inventories, POS, operations, services, etc.). Refactor AppSidebar into a role-aware navigation (NAVIGATION_ITEMS) with loading state and conditional NavUser rendering. Add session enforcement (requireStaff/requireAdmin) and tighten error handling/notFound behavior on several pages. Replace some <img> usages with next/Image and apply various UI/layout/CSS refinements. Add proxy middleware (proxy.ts) to redirect between /staff-login and /back-office based on session cookie.
