+++ app/(back-office)/back-office/inventories/page.tsx
