+++ app/(back-office)/back-office/inventory-categories/page.tsx
