+++ app/(back-office)/back-office/operations/[appointmentId]/[petId]/page.tsx
@@ -1,0 +1,35 @@
+import { getPetOperationDetail } from "@/modules/operation/queries/get-pet-operation-detail";
+import OperationDetailClient from "@/modules/operation/components/OperationDetailClient";
+import { notFound } from "next/navigation";
+import { SiteHeader } from "@/components/site-header";
+import BackButton from "@/components/BackButton";
+import { requireStaff } from "@/lib/session";
+
+interface PageProps {
+  params: Promise<{
+    appointmentId: string;
+    petId: string;
+  }>;
+}
+
+export default async function OperationDetailPage({ params }: PageProps) {
+  await requireStaff();
+
+  const { appointmentId, petId } = await params;
+  const result = await getPetOperationDetail(appointmentId, petId);
+
+  if (!result.success || !result.data) {
+    notFound();
+  }
+
+  return (
+    <>
+      <SiteHeader title="รายละเอียดคิวงาน" />
+      <div className="p-6">
+        <BackButton className="mb-4" />
+        {/* โยนข้อมูลที่ Merge แล้วไปยัง Client Component */}
+        <OperationDetailClient initialData={result.data} />
+      </div>
+    </>
+  );
+}
