+++ app/(back-office)/back-office/pos/[id]/page.tsx
