--- app/(back-office)/operations/[appointmentId]/[petId]/page.tsx
@@ -1,38 +1,0 @@
-import { getPetOperationDetail } from "@/modules/operation/queries/get-pet-operation-detail";
-import OperationDetailClient from "@/modules/operation/components/OperationDetailClient";
-import { notFound } from "next/navigation";
-import Link from "next/link";
-import { ChevronLeft } from "lucide-react";
-import { Button } from "@/components/ui/button";
-import { SiteHeader } from "@/components/site-header";
-import BackButton from "@/components/BackButton";
-import { requireStaff } from "@/lib/session";
-
-interface PageProps {
-  params: Promise<{
-    appointmentId: string;
-    petId: string;
-  }>;
-}
-
-export default async function OperationDetailPage({ params }: PageProps) {
-  await requireStaff();
-
-  const { appointmentId, petId } = await params;
-  const result = await getPetOperationDetail(appointmentId, petId);
-
-  if (!result.success || !result.data) {
-    notFound();
-  }
-
-  return (
-    <>
-      <SiteHeader title="รายละเอียดคิวงาน" />
-      <div className="p-6">
-        <BackButton className="mb-4" />
-        {/* โยนข้อมูลที่ Merge แล้วไปยัง Client Component */}
-        <OperationDetailClient initialData={result.data} />
-      </div>
-    </>
-  );
-}
